Update patch set 5
Patch Set 5:

(14 comments)

With Patchset #2 I notice that in testing locally there is no commit being added to the notedb repo, afaict: what am I missing to make that commit?

As noted in the commit message for patchset #2: this requires changing the visibility of the com.google.gerrit.server.account.externalids.ExternalIdCache interface in core.

Patch-set: 5